First version of the UPBReader with some additional functions.
Added a new dialog, to connect to a server dummy. Changed the longpress function. It is possible to add a note to a the selected text. And some additional changes to the code.

+++ src/org/geometerplus/fbreader/bookmodel/BookModel.java
@@ -143,11 +143,11 @@ public final class BookModel {
 		for (int i = 0; i < size; ++i) {
 			final char[] block = myInternalHyperlinks.block(i);
 			for (int offset = 0; offset < block.length; ) {
-				final int labelLength = (int)block[offset++];
+				final int labelLength = block[offset++];
 				if (labelLength == 0) {
 					break;
 				}
-				final int idLength = (int)block[offset + labelLength];
+				final int idLength = block[offset + labelLength];
 				if ((labelLength != len) || !id.equals(new String(block, offset, labelLength))) {
 					offset += labelLength + idLength + 3;
 					continue;
@@ -155,7 +155,7 @@ public final class BookModel {
 				offset += labelLength + 1;
 				final String modelId = (idLength > 0) ? new String(block, offset, idLength) : null;
 				offset += idLength;
-				final int paragraphNumber = (((int)block[offset++]) << 16) + (int)block[offset];
+				final int paragraphNumber = ((block[offset++]) << 16) + block[offset];
 				return new Label(modelId, paragraphNumber);
 			}
 		}
